docker-images/shinobi-smtp-proxy/requirements.txt
aiohttp
aiosmtpd
httpx
numpy
opencv-python-headless
pillow
prometheus-client
pyyaml
ultralytics
